Revert "Lineinfile"
